chore: Reuse test values between PreviewParameter and TestParameter
