+++ android/feature/nowplaying/src/test/snapshots/images/ly.david.musicsearch.android.feature.nowplaying_NowPlayingHistoryCardTest_default[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:fff836bc6efc09e2ad97979580d5137cb5a15f08912ca21f523c3904f80aba93
-size 16272
+oid sha256:abf70da5f097dbf77388e9065460e5a2944ce3ed68f756e95eb97e9327a98313
+size 16147
