+++ android/feature/nowplaying/src/test/snapshots/images/ly.david.musicsearch.android.feature.nowplaying_NowPlayingHistoryUiTest_default[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:813f696d9a030890de63c10822d0461ecc915cb318dca27ec27bc3874336a47a
-size 28459
+oid sha256:0c166bfceaae5f1d1479794ac00c880a5c6ac227674151b3e56654812ad81249
+size 28295
