+++ android/feature/nowplaying/src/test/snapshots/images/ly.david.musicsearch.android.feature.nowplaying_NowPlayingHistoryUiTest_default[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2d84474633dcd4e8f2be3a7567660f7157354ce19e6537a6875421d09aa276ca
-size 28917
+oid sha256:ec2f84e990353ac89c6190603fe2e089894b5ec4ccf815f902b752a8684af73f
+size 28761
