+++ android/feature/spotify/src/test/snapshots/images/ly.david.musicsearch.android.feature.spotify.history_SpotifyHistoryCardTest_default[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4d311ed0e6ce11fa16c7fcaaba5d4ef354fd184af3746e987c5c075d3bd73457
-size 17520
+oid sha256:abaabbf33cb25255286caa834dea862e85c68275fd58abcc370775e621e3facf
+size 17680
