+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artistWithCoverArt[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:4b5770ffa687e2d5aa019e2bbc45182c472bc4519b672bc476ef929de9d49854
-size 15973
+oid sha256:335910f20fbbe48caae39b2f7522dd77db870396bd76fb0fdd53f0b0bf4326dc
+size 16120
