+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_artist[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:d2659c344732967eab1bed441acb4c3e8934293b2870adc3bb77bcbac25e8184
-size 14235
+oid sha256:777df7e33eb071b415d978358589b5911db786d1544de06712e3b73a57eef0dd
+size 14384
