+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroupWithCoverArt[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1910e67e69d913a9bfc981ae2f61b78471aad2aa7b50ab0e979fc8c0ebd18117
-size 25791
+oid sha256:3ca71b95f0625ac79ad87306fd01cfc837991f99e88def9f11faa4e0f81a5894
+size 25843
