+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroup[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:05b97d390dffe8686a5d8693266d0b644cd73838427dc2a91fde1bc604310e54
-size 26552
+oid sha256:f44e61cc65134bc81db61ed0b9dd1799fd7a1fd245ffa39b78943adb9984024a
+size 26594
