+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseGroup[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:2413c39df7a1b5d70f124ea46fc5baec0e9f82c637f90002992138af7a11c7be
-size 26533
+oid sha256:67864ae5f54357c64613c1574bc4c9c394015ba0ae06c6bc491b55cae576b3fc
+size 26630
