+++ shared/feature/history/src/test/snapshots/images/ly.david.musicsearch.shared.feature.history_HistoryListItemTest_releaseWithCoverArt[NIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:f3ca1293674d9920121e88accbc4765b6f54f758bc2586f44817d570dfbcae03
-size 21746
+oid sha256:b865415599081e4f3d4688c656f18d1b7980f722c0e9c449ce520c30ac4a6643
+size 21796
