+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NIGHT,AreaListItemModel(id=1, name=Area Name, sortName=, disambiguation=, type=, lifeSpan=null, countryCodes=null, date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:78016975d1d258f418eab4e6e8d9d4621ec65dedaeb4b2e61d8231c38122deb4
+size 6588
