+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NIGHT,AreaListItemModel(id=2, name=Area Name, sortName=, disambiguation=That one, type=, lifeSpan=null, countryCodes=null, date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:23978425c7952034a607b5a5a52de324ff8c9bcfb93dddaf7c7c4b25e22ac0dc
+size 11302
