+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NIGHT,AreaListItemModel(id=3, name=Area Name with a very long name, sortName=, disambiguation=That one, type=Country, lifeSpan=null, countryCodes=[GB], date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a19d4ec71df28bcec6bc030e79c3cfeab2a061c70ef9ccf7c7721e1a4ae2fbce
+size 27837
