+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NIGHT,AreaListItemModel(id=4, name=Area Name with a very long name, sortName=, disambiguation=, type=[Worldwide], lifeSpan=null, countryCodes=[XW], date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b7fae2bd08dcb5a3d7130ddee43134ba3c262313a132dc6b19af7af966c27e87
+size 22376
