+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NOTNIGHT,AreaListItemModel(id=1, name=Area Name, sortName=, disambiguation=, type=, lifeSpan=null, countryCodes=null, date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:eb58ee84dbe87bc8baccaaed7a8339275b1388f640b060a94a3e770ed0975333
+size 7065
