+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NOTNIGHT,AreaListItemModel(id=2, name=Area Name, sortName=, disambiguation=That one, type=, lifeSpan=null, countryCodes=null, date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:5a60c48f29ca65c47a2b3f9ab472990469b9126abd70acee3971831bae7d21e4
+size 11724
