+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NOTNIGHT,AreaListItemModel(id=3, name=Area Name with a very long name, sortName=, disambiguation=That one, type=Country, lifeSpan=null, countryCodes=[GB], date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a0a50162ed1b46025c04982d8ff7c77277adbb700c3b76782db380d0f6d757b1
+size 28598
