+++ ui/common/src/test/snapshots/images/ly.david.ui.common.area_AreaListItemTest_areaListItem[NOTNIGHT,AreaListItemModel(id=4, name=Area Name with a very long name, sortName=, disambiguation=, type=[Worldwide], lifeSpan=null, countryCodes=[XW], date=null)].png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:c71942577bdec89d3de780be0b639b2a2994c25e5d1c84e37e5857bfe91189ec
+size 22968
