+++ ui/common/src/test/snapshots/images/ly.david.ui.common.artist_ArtistListItemTest_allInfo[NOTNIGHT].png
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1a72b6c058a5a2cfd662cb18744e8fb22da70b42b4b448ffc4303d5462dbe615
-size 82350
+oid sha256:9d86af47719664ec28bf59287933390a3fec7de8134675ede0d1e2ac9ea6c3a1
+size 82373
